backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

pydantic==2.6.1

transformers==4.38.2
torch==2.2.1

sentencepiece==0.1.99
accelerate==0.27.2

nltk==3.8.1
regex==2023.12.25

langdetect==1.0.9
deep-translator==1.11.4

pandas==2.2.1
numpy==1.26.4

reportlab==4.1.0
openpyxl==3.1.2
pymongo
bcrypt==4.2.1
python-dotenv==1.0.1
email-validator==2.1.1
dnspython==2.6.1
matplotlib

python-multipart==0.0.9
PyJWT==2.10.1
